+++ MyApp/package.json
@@ -20,6 +20,7 @@
     "@react-navigation/stack": "^6.1.1",
     "react": "17.0.2",
     "react-native": "0.67.3",
+    "react-native-date-picker": "^4.2.0",
     "react-native-gesture-handler": "^2.3.2",
     "react-native-pager-view": "^5.4.15",
     "react-native-paper": "^4.11.2",
